Add standalone minified file for some use cases
Use cases like including this file from a script tag, adding the name
UniversalEvents to the global namespace.

I've removed the dependency on async, as I only needed it for one thing
and it makes the packaged file larger.

Still depends on q for promises, might try to use the ES6 ones but that
will require a polyfill on browsers.

Packaging is done with browserify in a gulp task. Uses babelify and
uglify.