Raising an event makes a local copy of handlers
This is important, as some handlers actually modify the handlers for the
same event. The natural examples are cbOnce and await, which remove
their own handlers.

This didn't work before, which meant that awaiting the same event
multiple times did not work. There is now a test for this, which passes
after these changes.